+++ nodes/uibuilder.js
@@ -20,7 +20,7 @@
 //#region --- Type Defs --- //
 /**
  * @typedef {import('../typedefs.js')}
- * @typedef {import('node-red')} Red
+ * typedef {import('node-red')} Red
  */
 //#endregion --- Type Defs --- //
 
@@ -54,7 +54,7 @@ const uib = {
     nodeRoot: '',
     /** Track across redeployments @constant {Object} uib.deployments */
     deployments: {},
-    /** When nodeGo is run, add the node.id as a key with the value being the url
+    /** When nodeInstance is run, add the node.id as a key with the value being the url
      *  then add processing to ensure that the URL's are unique. 
      * Schema: {'<node.id>': '<url>'}
      * @constant {Object} uib.uib.instances
@@ -105,6 +105,18 @@ const uib = {
     staticOpts: {}, //{ maxAge: 31536000, immutable: true, },
     /** Array of instances that have requested their local instance folders be deleted on deploy - see html file oneditdelete, updated by admin api */
     deleteOnDelete: {},
+    /** Parameters for custom webserver if required. Port is undefined if using Node-RED's webserver. */
+    customServer: {
+        /** Optional TCP/IP port number. If defined, uibuilder will use its own ExpressJS server/app
+         * If undefined, uibuilder will use the Node-RED user-facing ExpressJS server
+         * @type {undefined|number} If undefined, means that uibuilder is using Node-RED's webserver
+         */
+        port: undefined,
+        /** @type {string} Node.js server type. ['http', 'https', 'http2']  */
+        type: 'http',
+        /** @type {undefined|string} uibuilder Host. sub(domain) name or IP Address */
+        host: undefined,
+    },
 }
 
 /** Current module version (taken from package.json) @constant {string} uib.version */
@@ -132,19 +144,8 @@ var commonStaticLoaded = false
 
 /** Export the function that defines the node 
  * @type Red */
-module.exports = function(/** @type Red */ RED) {
-    // NB: entries in settings.js are read-only and shouldn't be read using RED.settings.get, that is only for settings that can change in-flight.
-    //     see Node-RED issue #1543.
-
-    /*  RED.events.on('nodes-started',function() {
-            console.log('****** All nodes have started ******')
-        })
-        RED.events.on('nodes-stopped',function() {
-            console.log('****** All nodes have stopped ******')
-        }) 
-    */
-
-    //#region ----- Constants (& logging) for standard setup ----- //
+module.exports = function(/** @type {runtimeRED} */ RED) {
+    //#region ----- Constants for standard setup ----- //
     /** Folder containing settings.js, installed nodes, etc. @constant {string} userDir */
     // @ts-ignore
     userDir = RED.settings.userDir
@@ -161,9 +162,7 @@ module.exports = function(/** @type Red */ RED) {
     uib.configFolder = path.join(uib.rootFolder,uib.configFolderName) 
     uib.commonFolder = path.join(uib.rootFolder, uib.commonFolderName)
     
-    /** Root URL path for http-in/out and uibuilder nodes @constant {string} httpNodeRoot */
-    // @ts-ignore
-    const httpNodeRoot = uib.nodeRoot = RED.settings.httpNodeRoot
+    //#endregion -------- Constants -------- //
 
     //#region ----- back-end debugging ----- //
     // @ts-ignore
@@ -171,14 +170,67 @@ module.exports = function(/** @type Red */ RED) {
     log.trace('[uibuilder:Module] ----------------- uibuilder - module started -----------------')
     //#endregion ----- back-end debugging ----- //
 
+    //#region ----- ExpressJS Server/app setup ----- //
+
+    /** Optional port. If set, uibuilder will use its own ExpressJS server */
+    if ( RED.settings.uibuilder && RED.settings.uibuilder.port && RED.settings.uibuilder.port !== RED.settings.uiPort) 
+        uib.customServer.port = RED.settings.uibuilder.port
+
+    /** Root URL path for http-in/out and uibuilder nodes 
+     * Always set to empty string if a dedicated ExpressJS app is required
+     * @type {string} httpNodeRoot
+     */
+    let httpNodeRoot
+
     /** We need an http server to serve the page and vendor packages. 
      * @since 2019-02-04 removed httpAdmin - we only want to use httpNode for web pages 
-     * @since v2.0.0 2019-02-23 Moved from instance level (nodeGo()) to module level */
-    const app = RED.httpNode // || RED.httpAdmin
+     * @since v2.0.0 2019-02-23 Moved from instance level (nodeInstance()) to module level
+     * @since v3.3.0 2021-03-16 Allow independent ExpressJS server/app 
+     */
+    let app, server
+    if ( uib.customServer.port ) {
+        // Port has been specified & is different to NR's port so create a new instance of express & app
+        const express = require('express') 
+        app = express()
+        /** Socket.io needs an http(s) server rather than an ExpressJS app
+         * As we want Socket.io on the same port, we have to create out own server
+         * Use https if NR itself is doing so, use same certs as NR
+         * TODO: Allow for https/settings overrides using uibuilder props in settings.js
+         * TODO: Switch from https to http/2?
+         */
+        if ( RED.settings.https ) {
+            uib.customServer.type = 'https'
+            server = require('https').createServer(RED.settings.https, app)
+        } else {
+            server = require('http').createServer(app)
+        }
+        // Connect the server to the requested port, domain is the same as Node-RED
+        server.on('error', (err) => {
+            if (err.code === 'EADDRINUSE') {
+                server.close()
+                RED.log.error(
+                    `[uibuilder:CreateServer] ERROR: Port ${uib.customServer.port} is already in use. Cannot create uibuilder server, use a different port number and restart Node-RED`
+                )
+                return
+            }    
+        })
+        server.listen(uib.customServer.port, function() {
+            uib.customServer.host = server.address().address
+        })
+        // Override the httpNodeRoot setting, has to be empty string. Use reverse proxy to change.
+        httpNodeRoot = uib.nodeRoot = ''
+    } else {
+        // Port not specified (default) so reuse Node-RED's ExpressJS server and app
+        app = RED.httpNode // || RED.httpAdmin
+        server = RED.server
+        // Record the httpNodeRoot for later use
+        httpNodeRoot = uib.nodeRoot = RED.settings.httpNodeRoot
+    }
 
-    //#endregion -------- Constants -------- //
+    //#endregion ----- End of ExpressJS server/app setup ----- //
 
     //#region ----- Set up uibuilder root, root/.config & root/common folders ----- //
+
     /** Check uib root folder: create if needed, writable?
      * @since v2.0.0 2019-03-03
      */
@@ -220,17 +272,18 @@ module.exports = function(/** @type Red */ RED) {
             })
         } catch (e) {
             // should never happen
-            log.error(`[uibuilder] COPY OF COMMON FOLDER FAILED`)
+            log.error('[uibuilder] COPY OF COMMON FOLDER FAILED')
         }
-        // It is served up at the instance level to allow caching to be configured. It is used as a static resource folder (added in nodeGo() so available for each instance as `./common/`)
+        // It is served up at the instance level to allow caching to be configured. It is used as a static resource folder (added in nodeInstance() so available for each instance as `./common/`)
     }
     // If the root folder setup failed, throw an error and give up completely
     if (uib_rootFolder_OK !== true) {
         throw new Error(`uibuilder: Failed to set up uibuilder root folder structure correctly. Check log for additional error messages. Root folder: ${uib.rootFolder}.`)
     }
+
     //#endregion ----- root folder ----- //
     
-    /** Serve up vendor packages - updates uib.installedPackages
+    /** Serve up vendor packages. Updates uib.installedPackages
      * This is the first check, the installed packages are rechecked at various times.
      * Reads the packageList and masterPackageList files
      * Adds ExpressJS static paths for each found FE package & saves the details to the vendorPaths variable.
@@ -245,9 +298,7 @@ module.exports = function(/** @type Red */ RED) {
      **/
 
     /** URI path for accessing the socket.io client from FE code. 
-     * @since v2.0.0-dev Now: `../uibuilder/vendor/socket.io/socket.io.js`
      * @constant {string} uib_socketPath */
-    //const uib_socketPath = tilib.urlJoin(httpNodeRoot, uib.moduleName, 'socket.io')
     const uib_socketPath = tilib.urlJoin(httpNodeRoot, uib.moduleName, 'vendor', 'socket.io')
 
     log.trace('[uibuilder:Module] Socket.IO initialisation - Socket Path=', uib_socketPath )
@@ -257,16 +308,16 @@ module.exports = function(/** @type Red */ RED) {
         // Allow-Headers:X-ClientId in there.  see https://socket.io/docs/v2/handling-cors/
         handlePreflightRequest: (req, res) => {
             res.writeHead(204, {
-              "Access-Control-Allow-Origin": req.headers["origin"],
-              "Access-Control-Allow-Methods": "GET,POST",
-              "Access-Control-Allow-Headers": "X-ClientId",
-              "Access-Control-Allow-Credentials": true,
-            });
-            res.end();
+                'Access-Control-Allow-Origin': req.headers['origin'],
+                'Access-Control-Allow-Methods': 'GET,POST',
+                'Access-Control-Allow-Headers': 'X-ClientId',
+                'Access-Control-Allow-Credentials': true,
+            })
+            res.end()
         },
     }
-    var io = socketio.listen(RED.server, ioOptions) // listen === attach
-    // @ts-ignore
+    var io = socketio.listen(server, ioOptions) // listen === attach
+    // @ts-expect-error ts(2339)
     io.set('transports', ['polling', 'websocket'])
 
     /** Check for <uibRoot>/.config/sioMiddleware.js, use it if present. Copy template if not exists @since v2.0.0-dev3 */
@@ -282,10 +333,10 @@ module.exports = function(/** @type Red */ RED) {
 
     //#endregion ----- socket.io server ----- //
 
-    //#region ---- Set up uibuilder master resources (these are applied in nodeGO at instance level) ----
+    //#region ---- Set up uibuilder master resources (these are applied in nodeInstance at instance level) ----
     /** Create a new, additional static http path to enable loading of central static resources for uibuilder
      * Loads standard images, ico file, etc.
-     * @since v2.0.0 2019-03-03 Moved out of nodeGo() only need to do once - but is applied in nodeGo
+     * @since v2.0.0 2019-03-03 Moved out of nodeInstance() only need to do once - but is applied in nodeInstance
      * @type {string}
      */
     var masterStatic = ''
@@ -306,21 +357,25 @@ module.exports = function(/** @type Red */ RED) {
     // the default index.html page can be utilised.
     //#endregion -------- master resources --------
 
-    { // Output startup info to Node-RED log
-        RED.log.info('+-----------------------------------------------------')
-        RED.log.info(`| ${uib.moduleName} initialised:`)
-        RED.log.info(`|   root folder: ${uib.rootFolder}`)
-        RED.log.info(`|   version . .: ${uib.version}`)
-        RED.log.info(`|   packages . : ${Object.keys(uib.installedPackages)}`)
-        RED.log.info('+-----------------------------------------------------')
-    }
+    //#region ---- Output startup info to Node-RED log ---- //
+    RED.log.info('+-----------------------------------------------------')
+    RED.log.info(`| ${uib.moduleName} initialised:`)
+    if ( uib.customServer.port )
+        RED.log.info(`|   Using custom ${uib.customServer.type} webserver on port ${uib.customServer.port}`)
+    else
+        RED.log.info('|   Using Node-RED\'s webserver')
+    RED.log.info(`|   root folder: ${uib.rootFolder}`)
+    RED.log.info(`|   version . .: ${uib.version}`)
+    RED.log.info(`|   packages . : ${Object.keys(uib.installedPackages)}`)
+    RED.log.info('+-----------------------------------------------------')
+    //#endregion ------------------------------------------ //
 
     /** Run the node instance - called from registerType()
-     * @param {Object} config The configuration object passed from the Admin interface (see the matching HTML file)
+     * @type {runtimeNode}
+     * @param {runtimeNodeConfig} config The configuration object passed from the Admin interface (see the matching HTML file)
      */
-    function nodeGo(config) {
+    function nodeInstance(config) {
         // Create the node
-        // @ts-expect-error ts(2339)
         RED.nodes.createNode(this, config)
 
         /** @since 2019-02-02 - the current instance name (url) */
@@ -328,10 +383,9 @@ module.exports = function(/** @type Red */ RED) {
 
         log.trace(`[uibuilder:${uibInstance}] ================ instance registered ================`)
 
-        /** A copy of 'this' object in case we need it in context of callbacks of other functions. 
-         * @type {uibNode} node
+        /** Copy 'this' object in case we need it in context of callbacks of other functions.
+         * @type {uibNode}
          */
-        // @ts-expect-error ts(2322)
         const node = this
         log.trace(`[uibuilder:${uibInstance}] = Keys: this, config =`, {'this': Object.keys(node), 'config': Object.keys(config)})
 
@@ -349,8 +403,7 @@ module.exports = function(/** @type Red */ RED) {
         node.showfolder      = config.showfolder === undefined ? false : config.showfolder
         node.useSecurity     = config.useSecurity 
         node.sessionLength   = Number(config.sessionLength) || 120  // in seconds
-        // @ts-expect-error ts(2339)
-        node.jwtSecret       = this.credentials.jwtSecret || 'thisneedsreplacingwithacredential'
+        node.jwtSecret       = node.credentials.jwtSecret || 'thisneedsreplacingwithacredential'
         node.tokenAutoExtend = config.tokenAutoExtend
         //#endregion ----- Local node config copy ----- //
 
@@ -439,7 +492,7 @@ module.exports = function(/** @type Red */ RED) {
         }
 
         //#region Add static ExpressJS route for instance local custom files
-        var customStatic = function(req, res, next) { next() } // Dummy ExpressJS middleware, replaced by local static folder if needed
+        var customStatic = function(/** @type {express.Request} */ req, /** @type {express.Response} */ res, /** @type {express.NextFunction} */ next) { next() } // Dummy ExpressJS middleware, replaced by local static folder if needed
 
         try {
             // Check if local dist folder contains an index.html & if NR can read it - fall through to catch if not
@@ -477,7 +530,7 @@ module.exports = function(/** @type Red */ RED) {
         /** Provide the ability to have a ExpressJS middleware hook.
          * This can be used for custom authentication/authorisation or anything else.
          */
-        var httpMiddleware = function(req,res,next) { next() }
+        var httpMiddleware = function(/** @type {express.Request} */ req, /** @type {express.Response} */ res, /** @type {express.NextFunction} */ next) { next() }
         /** Check for <uibRoot>/.config/uibMiddleware.js, use it if present. Copy template if not exists @since v2.0.0-dev4 */
         let uibMwPath = path.join(uib.configFolder, 'uibMiddleware.js')
         try {
@@ -491,7 +544,7 @@ module.exports = function(/** @type Red */ RED) {
 
         /** This ExpressJS middleware runs when the uibuilder page loads - set cookies and headers
          * @see https://expressjs.com/en/guide/using-middleware.html */
-        function masterMiddleware (req, res, next) {
+        function masterMiddleware (/** @type {express.Request} */ req, /** @type {express.Response} */ res, /** @type {express.NextFunction} */ next) {
             //TODO: X-XSS-Protection only needed for html (and js?), not for css, etc
             // Help reduce risk of XSS and other attacks
             res.setHeader('X-XSS-Protection','1;mode=block')
@@ -711,7 +764,7 @@ module.exports = function(/** @type Red */ RED) {
          * @param {function} done Per msg finish function, node-red v1+
          **/
         function nodeInputHandler(msg, send, done) {
-            log.trace(`[uibuilder:${uibInstance}] nodeGo:nodeInputHandler - emit received msg - Namespace: ${node.url}`) //debug
+            log.trace(`[uibuilder:${uibInstance}] nodeInstance:nodeInputHandler - emit received msg - Namespace: ${node.url}`) //debug
 
             // If this is pre-1.0, 'send' will be undefined, so fallback to node.send
             send = send || function() { node.send.apply(node,arguments) }
@@ -743,7 +796,7 @@ module.exports = function(/** @type Red */ RED) {
 
         // Do something when Node-RED is closing down which includes when this node instance is redeployed
         node.on('close', function(removed,done) {
-            log.trace(`[uibuilder:${uibInstance}] nodeGo:on-close: ${removed?'Node Removed':'Node (re)deployed'}`)
+            log.trace(`[uibuilder:${uibInstance}] nodeInstance:on-close: ${removed?'Node Removed':'Node (re)deployed'}`)
 
             node.removeListener('input', nodeInputHandler)
 
@@ -755,23 +808,24 @@ module.exports = function(/** @type Red */ RED) {
         })
 
         // Shows an instance details debug page
-        RED.httpAdmin.get(`/uibuilder/instance/${node.url}`, function(req,res) {
-            let page = uiblib.showInstanceDetails(node, uib, userDir, RED)
+        RED.httpAdmin.get(`/uibuilder/instance/${node.url}`, function(/** @type {express.Request} */ req, /** @type {express.Response} */ res) {
+            let page = uiblib.showInstanceDetails(req, node, uib, userDir, RED)
             res.status(200).send( page )
         })
 
-    } // ---- End of nodeGo (initialised node instance) ---- //
+    } // ---- End of nodeInstance (initialised node instance) ---- //
 
     /** Register the node by name. This must be called before overriding any of the
      *  Node functions. */
-    // @ts-ignore
-    RED.nodes.registerType(uib.moduleName, nodeGo, {
+    console.log('===========', uib.customServer, '==============')
+    RED.nodes.registerType(uib.moduleName, nodeInstance, {
         credentials: {
             jwtSecret: {type:'password'},
         },
         settings: {
             uibuilderNodeEnv: { value: process.env.NODE_ENV, exportable: true },
             uibuilderTemplates: { value: templateConf, exportable: true },
+            uibuilderCustomServer: { value: (uib.customServer), exportable: true },
         },
     })
 
@@ -893,7 +947,7 @@ module.exports = function(/** @type Red */ RED) {
     /** uibuilder v3 unified Admin API router - new API commands should be added here */
     RED.httpAdmin.route('/uibuilder/admin/:url')
         // For all routes
-        .all(function(req,res,next) {
+        .all(function(/** @type {express.Request} */ req, /** @type {express.Response} */ res, /** @type {express.NextFunction} */ next) {
             // @ts-ignore
             const params = res.allparams = Object.assign({}, req.query, req.body, req.params)
             params.type = 'all'
@@ -911,7 +965,7 @@ module.exports = function(/** @type Red */ RED) {
             next()
         })
         /** Get something and return it */
-        .get(function(req,res) {
+        .get(function(/** @type {express.Request} */ req, /** @type {express.Response} */ res) {
             // @ts-ignore
             const params = res.allparams
             params.type = 'get'
@@ -976,7 +1030,7 @@ module.exports = function(/** @type Red */ RED) {
                 })
                 console.log(app._router.stack[0])
 
-                log.trace(`[uibuilder:admin-router:GET:listurls] Admin API. List of all user urls in use.`)
+                log.trace('[uibuilder:admin-router:GET:listurls] Admin API. List of all user urls in use.')
                 res.statusMessage = 'URLs listed successfully'
                 //res.status(200).json(routes)
                 res.status(200).json(app._router.stack)
@@ -986,7 +1040,7 @@ module.exports = function(/** @type Red */ RED) {
 
         })
         /** TODO Write file contents */
-        .put(function(req,res) {
+        .put(function(/** @type {express.Request} */ req, /** @type {express.Response} */ res) {
             // @ts-ignore
             const params = res.allparams
             params.type = 'put'
@@ -1011,7 +1065,7 @@ module.exports = function(/** @type Red */ RED) {
 
         })
         /** Create a new folder or file */
-        .post(function(req,res) {
+        .post(function(/** @type {express.Request} */ req, /** @type {express.Response} */ res) {
             // @ts-ignore
             const params = res.allparams
             params.type = 'post'
@@ -1080,7 +1134,7 @@ module.exports = function(/** @type Red */ RED) {
             })
         })
         /** Delete a folder or a file */
-        .delete(function(req,res) {
+        .delete(function(/** @type {express.Request} */ req, /** @type {express.Response} */ res) {
             // @ts-ignore ts(2339)
             const params = res.allparams
             params.type = 'delete'
@@ -1159,7 +1213,7 @@ module.exports = function(/** @type Red */ RED) {
      * @param {Object} permissions The permissions required for access
      * @param {function} cb
      **/
-    RED.httpAdmin.get('/uibgetfile', function(req,res) {
+    RED.httpAdmin.get('/uibgetfile', function(/** @type {express.Request} */ req, /** @type {express.Response} */ res) {
         //#region --- Parameter validation ---
         /** req.query parameters
          * url
@@ -1233,7 +1287,7 @@ module.exports = function(/** @type Red */ RED) {
      * @param {Object} permissions The permissions required for access (Express middleware)
      * @param {function} cb
      **/
-    RED.httpAdmin.post('/uibputfile', function(req,res) {
+    RED.httpAdmin.post('/uibputfile', function(/** @type {express.Request} */ req, /** @type {express.Response} */ res) {
         //#region --- Parameter validation ---
         const params = req.body
 
@@ -1286,8 +1340,16 @@ module.exports = function(/** @type Red */ RED) {
      * Also allows confirmation of whether a url is in use ('check' parameter) or a simple list of urls in use.
      * @since 2019-02-04 v1.1.0-beta6
      */
-    RED.httpAdmin.get('/uibindex', function(req,res) {
+    RED.httpAdmin.get('/uibindex', function(/** @type {express.Request} */ req, /** @type {express.Response} */ res) {
         log.trace('[uibindex] User Page/API. List all available uibuilder endpoints')
+        
+        // If using own Express server, correct the URL's
+        const url = new URL(req.headers.referer)
+        url.pathname = ''
+        if (uib.customServer.port) {
+            url.port = uib.customServer.port
+        }
+        const urlPrefix = url.href
         
         /** Return full details based on type parameter */
         switch (req.query.type) {
@@ -1349,8 +1411,8 @@ module.exports = function(/** @type Red */ RED) {
                 let page = `
                     <!doctype html><html lang="en"><head>
                         <title>Uibuilder Index</title>
-                        <link type="text/css" href="${uib.nodeRoot}${uib.moduleName}/vendor/bootstrap/dist/css/bootstrap.min.css" rel="stylesheet" media="screen">
-                        <link rel="icon" href="${uib.nodeRoot}${uib.moduleName}/common/images/node-blue.ico">
+                        <link type="text/css" href="${urlPrefix}${uib.nodeRoot.replace('/','')}${uib.moduleName}/vendor/bootstrap/dist/css/bootstrap.min.css" rel="stylesheet" media="screen">
+                        <link rel="icon" href="${urlPrefix}${uib.nodeRoot.replace('/','')}${uib.moduleName}/common/images/node-blue.ico">
                         <style type="text/css" media="all">
                             h2 { border-top:1px solid silver;margin-top:1em;padding-top:0.5em; }
                             .col3i tbody>tr>:nth-child(3){ font-style:italic; }
@@ -1376,7 +1438,7 @@ module.exports = function(/** @type Red */ RED) {
                 Object.keys(uib.instances).forEach(key => {
                     page += `
                         <tr>
-                            <td><a href="${tilib.urlJoin(httpNodeRoot, uib.instances[key])}" target="_blank">${uib.instances[key]}</a></td>
+                            <td><a href="${urlPrefix}${tilib.urlJoin(httpNodeRoot, uib.instances[key]).replace('/','')}" target="_blank">${uib.instances[key]}</a></td>
                             <td>${key}</td>
                             <td>${path.join(uib.rootFolder, uib.instances[key])}</td>
                         </tr>
@@ -1421,11 +1483,11 @@ module.exports = function(/** @type Red */ RED) {
                      *  If so, add them to the output as an indicator of where to look.
                      */
                     let mainTxt = '<i>Not Supplied</i>'
-                    console.log('==>> ',httpNodeRoot, pj.url,pj.browser)
+                    //console.log('==>> ',httpNodeRoot, pj.url,pj.browser)
                     if ( pj.browser !== '' ) {
-                        mainTxt = `<a href="${tilib.urlJoin(httpNodeRoot, pj.url.replace('..',''), pj.browser)}">${pj.url}/${pj.browser}</a>`
+                        mainTxt = `<a href="${urlPrefix}${tilib.urlJoin(httpNodeRoot, pj.url.replace('..',''), pj.browser).replace('/','')}">${pj.url}/${pj.browser}</a>`
                     } else if ( pj.main !== '' ) {
-                        mainTxt = `<a href="${tilib.urlJoin(httpNodeRoot, pj.url.replace('..',''), pj.main)}">${pj.url}/${pj.main}</a>`
+                        mainTxt = `<a href="${urlPrefix}${tilib.urlJoin(httpNodeRoot, pj.url.replace('..',''), pj.main).replace('/','')}">${pj.url}/${pj.main}</a>`
                     }
 
                     page += `
@@ -1596,7 +1658,7 @@ module.exports = function(/** @type Red */ RED) {
     }) // ---- End of uibindex ---- //
 
     /** Check & update installed front-end library packages, return list as JSON */
-    RED.httpAdmin.get('/uibvendorpackages', function(req,res) {
+    RED.httpAdmin.get('/uibvendorpackages', function(/** @type {express.Request} */ req, /** @type {express.Response} */ res) {
         // Update the installed packages list
         uiblib.checkInstalledPackages('', uib, userDir, log)
 
@@ -1611,7 +1673,7 @@ module.exports = function(/** @type Red */ RED) {
      * @param {string} [req.query.url=userDir] Optional. If present, CWD is set to the uibuilder folder for that instance. Otherwise CWD is set to the userDir.
      * @param {string} req.query.cmd Command to run (see notes for this function)
      */
-    RED.httpAdmin.get('/uibnpmmanage', function(req,res) {
+    RED.httpAdmin.get('/uibnpmmanage', function(/** @type {express.Request} */ req, /** @type {express.Response} */ res) {
         //#region --- Parameter validation (cmd, package) ---
         const params = req.query