+++ templates/.config/security.js
@@ -1,6 +1,6 @@
 /* globals module */
 /**
- * Copyright (c) 2020 Julian Knight (Totally Information)
+ * Copyright (c) 2020-2021 Julian Knight (Totally Information)
  * https://it.knightnet.org.uk
  *
  * Licensed under the Apache License, Version 2.0 (the 'License');
@@ -37,20 +37,15 @@
  */
 'use strict'
 
-/**
- * Standard msg._auth object exchanged in msg's between front-end and server
- * @typedef {import('../../index').MsgAuth} MsgAuth
- */ 
-/**
- * Validated user object returned by the userValidate function
- * typedef {import('./security').userValidation} userValidation 
- */
-
-const TYPEDEFS = require('../../typedefs.js')
-/**
- * typedef {TYPEDEFS.MsgAuth} MsgAuth
- * @typedef {TYPEDEFS.userValidation} userValidation
- * @typedef {TYPEDEFS.userMetadata} userMetadata
+/** Define the _auth object
+ * @typedef {Object} _auth The standard auth object used by uibuilder security. See docs for details.
+ * Note that any other data may be passed from your front-end code in the uibAuth object.
+ * @property {String} id Required. A unique user identifier.
+ * @property {String} [password] Required for input to login only.
+ * @property {String} [jwt] Required if logged in. Needed for ongoing session validation and management.
+ * @property {String} [sessionExpiry] Required if logged in. ISO8601 formatted date/time string. Needed for ongoing session validation and management.
+ * @property {boolean} [userValidated] Required after user validation. Whether the input ID (and optional additional data from the _auth object) validated correctly or not.
+ * @property {Object=} [info] Optional metadata about the user.
  */
 
 module.exports = {
@@ -58,42 +53,40 @@ module.exports = {
      * The minimum input data is _auth.id which must represent a unique ID.
      * Called from the logon function (uiblib.js::logon) which is triggered by a uibuilder control msg from the client of type 'logon'.
      * May also be called to revalidate users at any time.
-     * @param {MsgAuth} _auth Required. 
-     * @return {boolean|userValidation} Either true/false or Object of type userValidation
+     * @param {_auth} _auth Required. 
+     * @return {_auth} Object of type _auth
      */
     userValidate: function(_auth) {
-        console.log(`[uibuilder:security.js] userValidate Security from ${__filename} used. Replace this template with your own code. _auth:`, _auth)
+        console.log(`[uibuilder:.common/security.js] userValidate Security from '${__filename}' used. Replace this template with your own code. _auth:`, _auth)
 
+        // Always start by invalidating the user credentials
+        _auth.userValidated = false
+        
         /** Manual "test" ID validates - this will be replaced with a suitable lookup in your code - maybe from a database or a file.
          * You will also want to pass through some kind of password to validate the user.
          */
         if ( _auth.id === 'test' ) {
-            console.log(`[uibuilder:security.js] User id ${_auth.id} has been validated`)
 
-            // Example of simple boolean return
-            return true
+            console.log(`[uibuilder:.common/security.js] User id '${_auth.id}' has been validated`)
 
-            //Example of object return with additional data that gets passed back to the client
-            // return {
-            //     userValidated: true,
-            //     authData: {
-            //         name: 'Me',
-            //         message: 'Hi you, don\'t forget to change your password :)'
-            //     }
-            // }
+            _auth.userValidated = true
+            _auth.info = {
+                name: 'Jimbob',
+                message: 'Hi Jimbob, don\'t forget to change your password :)'
+            }
+
+        } else {
+
+            // In all other cases, fail the validation - optionally, you can include more info here as well.
+            _auth.userValidated = false
+            _auth.info = {
+                message: 'Ouch! Sorry, that login is not valid'
+            }
         }
-
-        // In all other cases, fail the validation - optionally, you can include more info here by returning an object.
-        return false
-        // return {
-        //     userValidated: false,
-        //     authData: {
-        //         message: `Login failed, User id ${_auth.id} not recognised.`
-        //     }
-        // }
         
-    } // ---- End of userValidate ---- //
+        return _auth
 
+    } // ---- End of userValidate ---- //
 
 }